+++ utils/sessionClaimValidatorStore.d.ts
@@ -12,6 +12,14 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-export const package_version = "0.4.0";
 
-export const supported_fdi = ["1.15"];
+export * from "../lib/build/sessionClaimValidatorStore";
+/**
+ * 'export *' does not re-export a default.
+ * import Querier from "supertokens-web-js/utils/querier";
+ * the above import statement won't be possible unless either
+ * - user add "esModuleInterop": true in their tsconfig.json file
+ * - we do the following change:
+ */
+import _default from "../lib/build/sessionClaimValidatorStore";
+export default _default;